Add some minor variacne to body part relative volume

+++ MagiRogue/Entities/BodyPart.cs
@@ -42,7 +42,7 @@ namespace MagiRogue.Entities
             }
         }
 
-        public double MaxBodyPartHp { get; set; }
+        public int MaxBodyPartHp { get; set; }
 
         public double BodyPartWeight
         {
